# services/book-catalog-service/src/main/resources/openapi/internal-api.yaml
openapi: 3.0.3
info:
  title: Book Catalog Service Internal API
  description: Internal API for inter-service communication (not exposed via API Gateway)
  version: 1.0.0
servers:
  - url: http://localhost:8091
    description: Local development

paths:
  /api/internal/books/{id}:
    get:
      summary: Get book by ID
      operationId: getBookById
      tags:
        - Internal Book API
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: Book found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookResponse"
        "404":
          description: Book not found

  /api/internal/books/validate/{id}:
    get:
      summary: Validate if book exists
      operationId: validateBook
      tags:
        - Internal Book API
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResponse"

  /api/internal/books/batch:
    post:
      summary: Get multiple books by IDs
      operationId: getBooksBatch
      tags:
        - Internal Book API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
                format: int32
      responses:
        "200":
          description: Books map (bookId -> BookResponse)
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: "#/components/schemas/BookResponse"

  /api/internal/books/{id}/rating-cache:
    put:
      summary: Update cached rating for a book
      description: Called by Comment Rating Service when rating is added/updated/deleted
      operationId: updateRatingCache
      tags:
        - Internal Book API
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RatingCacheUpdateRequest"
      responses:
        "200":
          description: Rating cache updated successfully
        "404":
          description: Book not found

components:
  schemas:
    BookResponse:
      type: object
      required:
        - id
        - title
      properties:
        id:
          type: integer
          format: int32
        title:
          type: string
        description:
          type: string
        authors:
          type: array
          items:
            type: string
        genres:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        language:
          type: string
        yearPublished:
          type: integer
        coverImagePath:
          type: string
        createdAt:
          type: string
          format: date-time
        isDeleted:
          type: boolean

    ValidationResponse:
      type: object
      required:
        - exists
      properties:
        exists:
          type: boolean
        message:
          type: string

    RatingCacheUpdateRequest:
      type: object
      required:
        - averageRating
        - ratingsCount
      properties:
        averageRating:
          type: number
          minimum: 0.0
          maximum: 5.0
          description: Average rating (0.00-5.00)
        ratingsCount:
          type: integer
          format: int32
          minimum: 0
          description: Total number of ratings